feat(EventModel): detach model from event decoder
