+++ EventModel/IMetadata.cs
@@ -1,0 +1,16 @@
+﻿// <copyright file="IMetadata.cs" company="PlaceholderCompany">
+// Copyright (c) PlaceholderCompany. All rights reserved.
+// </copyright>
+
+namespace Nanolite_agent.EventModel
+{
+    /// <summary>
+    /// Represents a contract for metadata information used internally within the application.
+    /// </summary>
+    /// <remarks>This interface is intended for internal use and provides a structure for defining
+    /// metadata-related functionality. Implementations of this interface should encapsulate metadata details relevant
+    /// to their specific context.</remarks>
+    public interface IMetadata
+    {
+    }
+}
