+++ Helper/SysmonEventDecoder.cs
@@ -6,824 +6,9 @@ namespace Nanolite_agent.Helper
 {
     using System;
     using Microsoft.Diagnostics.Tracing;
+    using Nanolite_agent.EventModel;
     using Nanolite_agent.NanoException;
-    using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
-
-    /// <summary>
-    /// Represents a contract for metadata information.
-    /// </summary>
-    /// <remarks>This interface serves as a marker for classes that provide metadata. Implementing this
-    /// interface indicates that the SysmonEvent class contains metadata-related information.</remarks>
-    public interface IMetadata { }
-
-    /// <summary>
-    /// Represents the base class for a Sysmon event, providing common properties for event data.
-    /// </summary>
-    /// <remarks>This class serves as a foundation for specific Sysmon event types, encapsulating shared
-    /// properties such as timestamp, event code, event name, and source. Derived classes should implement the  <see
-    /// cref="MetadataObject"/> property to provide event-specific metadata.</remarks>
-    public abstract class EventBase
-    {
-        /// <summary>
-        /// Gets or sets the timestamp of the Sysmon event.
-        /// </summary>
-        [JsonProperty(nameof(Timestamp))]
-        public DateTime Timestamp { get; set; }
-
-        /// <summary>
-        /// Gets or sets the event code associated with the Sysmon event.
-        /// </summary>
-        [JsonProperty(nameof(EventCode))]
-        public int EventCode { get; set; }
-
-        /// <summary>
-        /// Gets or sets the name of the Sysmon event.
-        /// </summary>
-        [JsonProperty(nameof(EventName))]
-        public string EventName { get; set; }
-
-        /// <summary>
-        /// Gets or sets the source of the Sysmon event.
-        /// </summary>
-        [JsonProperty(nameof(Source))]
-        public string Source { get; set; }
-
-        [JsonIgnore]
-        public abstract IMetadata MetadataObject { get; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with the creation of a process.
-    /// </summary>
-    /// <remarks>This class provides detailed information about a process, including its command line, 
-    /// associated company, current directory, file version, and more. It also includes metadata  about the parent
-    /// process and user context. This information can be useful for auditing,  monitoring, or analyzing process
-    /// creation events.</remarks>
-    public sealed class ProcessCreationMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the command line used to start the process.
-        /// </summary>
-        [JsonProperty(nameof(CommandLine))]
-        public string CommandLine { get; set; }
-
-        /// <summary>
-        /// Gets or sets the company associated with the process image.
-        /// </summary>
-        [JsonProperty(nameof(Company))]
-        public string Company { get; set; }
-
-        /// <summary>
-        /// Gets or sets the current directory of the process.
-        /// </summary>
-        [JsonProperty(nameof(CurrentDirectory))]
-        public string CurrentDirectory { get; set; }
-
-        /// <summary>
-        /// Gets or sets the file version of the process image.
-        /// </summary>
-        [JsonProperty(nameof(FileVersion))]
-        public string FileVersion { get; set; }
-
-        /// <summary>
-        /// Gets or sets the hash values of the process image.
-        /// </summary>
-        [JsonProperty(nameof(Hashes))]
-        public string Hashes { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the integrity level of the process.
-        /// </summary>
-        [JsonProperty(nameof(IntegrityLevel))]
-        public string IntegrityLevel { get; set; }
-
-        /// <summary>
-        /// Gets or sets the logon ID associated with the process.
-        /// </summary>
-        [JsonProperty(nameof(LogonId))]
-        public string LogonId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the original file name of the process image.
-        /// </summary>
-        [JsonProperty(nameof(OriginalFileName))]
-        public string OriginalFileName { get; set; }
-
-        /// <summary>
-        /// Gets or sets the command line of the parent process.
-        /// </summary>
-        [JsonProperty(nameof(ParentCommandLine))]
-        public string ParentCommandLine { get; set; }
-
-        /// <summary>
-        /// Gets or sets the parent image name associated with the network connection.
-        /// </summary>
-        [JsonProperty(nameof(ParentImage))]
-        public string ParentImage { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID of the parent process.
-        /// </summary>
-        [JsonProperty(nameof(ParentProcessId))]
-        public long ParentProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the product name associated with the process image.
-        /// </summary>
-        [JsonProperty(nameof(Product))]
-        public string Product { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user name.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata related to process access events, including information about the source and target
-    /// processes and users.
-    /// </summary>
-    /// <remarks>This class provides properties to access details about the source and target processes
-    /// involved in an access event, such as their image names, process IDs, and associated users.</remarks>
-    public sealed class ProcessAccessMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the image name of the source process involved in the access event.
-        /// </summary>
-        [JsonProperty(nameof(SourceImage))]
-        public string SourceImage { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID of the source process involved in the access event.
-        /// </summary>
-        [JsonProperty(nameof(SourceProcessId))]
-        public long SourceProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the source process.
-        /// </summary>
-        [JsonProperty(nameof(SourceUser))]
-        public string SourceUser { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name of the target process involved in the access event.
-        /// </summary>
-        [JsonProperty(nameof(TargetImage))]
-        public string TargetImage { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID of the target process involved in the access event.
-        /// </summary>
-        [JsonProperty(nameof(TargetProcessId))]
-        public long TargetProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the target process.
-        /// </summary>
-        [JsonProperty(nameof(TargetUser))]
-        public string TargetUser { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a terminated process.
-    /// </summary>
-    /// <remarks>This class provides information about a process that has been terminated, including its ID,
-    /// the image name, and the user who initiated it.</remarks>
-    public sealed class ProcessTerminatedMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the process ID associated with the DNS query event.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the DNS query event.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the event.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with the creation of a remote thread in a target process.
-    /// </summary>
-    public sealed class CreateRemoteThreadMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the process ID of the source process that created the remote thread.
-        /// </summary>
-        [JsonProperty(nameof(SourceProcessId))]
-        public long SourceProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name of the source process that created the remote thread.
-        /// </summary>
-        [JsonProperty(nameof(SourceImage))]
-        public string SourceImage { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the source process.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID of the target process where the remote thread was created.
-        /// </summary>
-        [JsonProperty(nameof(TargetProcessId))]
-        public long TargetProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name of the target process where the remote thread was created.
-        /// </summary>
-        [JsonProperty(nameof(TargetImage))]
-        public string TargetImage { get; set; }
-
-        /// <summary>
-        /// Gets or sets the module name where the remote thread was created.
-        /// </summary>
-        public string StartModule { get; set; }
-
-        /// <summary>
-        /// Gets or sets the function name where the remote thread was created.
-        /// </summary>
-        public string StartFunction { get; set; }
-    }
-
-    public sealed class ProcessTamperingMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the user associated with the source process.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the DNS query event.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the DNS query event.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the type of tampering that occurred.
-        /// </summary>
-        public string Type { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata for a network connection, providing details about the source and destination endpoints, the
-    /// protocol used, and the process initiating the connection.
-    /// </summary>
-    /// <remarks>This class is used to encapsulate information about a network connection, including both
-    /// source and destination details such as IP addresses, hostnames, and ports. It also includes metadata about the
-    /// process that initiated the connection, such as the process ID and GUID, as well as user and image information.
-    /// This metadata can be useful for logging, monitoring, or analyzing network activity.</remarks>
-    public sealed class NetworkConnectionMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the destination host name for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(DestinationHostname))]
-        public string DestinationHostname { get; set; }
-
-        /// <summary>
-        /// Gets or sets the destination IP address for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(DestinationIp))]
-        public string DestinationIp { get; set; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether the destination IP is IPv6.
-        /// </summary>
-        [JsonProperty(nameof(DestinationIsIpv6))]
-        public bool DestinationIsIpv6 { get; set; }
-
-        /// <summary>
-        /// Gets or sets the destination port for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(DestinationPort))]
-        public long DestinationPort { get; set; }
-
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether the connection was initiated.
-        /// </summary>
-        [JsonProperty(nameof(Initiated))]
-        public bool Initiated { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process GUID associated with the network connection.
-        /// </summary>
-        [JsonProperty(nameof(ProcessGuid))]
-        public string ProcessGuid { get; set; }
-
-        /// <summary>
-        /// Gets or sets the unique identifier for the process.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the protocol used for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(Protocol))]
-        public string Protocol { get; set; }
-
-        /// <summary>
-        /// Gets or sets the source host name for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(SourceHostname))]
-        public string SourceHostname { get; set; }
-
-        /// <summary>
-        /// Gets or sets the source IP address for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(SourceIp))]
-        public string SourceIp { get; set; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether the source IP is IPv6.
-        /// </summary>
-        [JsonProperty(nameof(SourceIsIpv6))]
-        public bool SourceIsIpv6 { get; set; }
-
-        /// <summary>
-        /// Gets or sets the source port for the network connection.
-        /// </summary>
-        [JsonProperty(nameof(SourcePort))]
-        public long SourcePort { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the current operation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the identifier of the parent image.
-        /// </summary>
-        [JsonProperty(nameof(ParentImage))]
-        public string ParentImage { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a DNS query, including details about the process and user that initiated the
-    /// query.
-    /// </summary>
-    /// <remarks>This class provides properties to access information such as the image name of the process,
-    /// the process ID, the DNS query name, and the user who initiated the query.</remarks>
-    public sealed class DnsQueryMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the URL of the image associated with this entity.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the unique identifier for the process.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the DNS query name.
-        /// </summary>
-        [JsonProperty(nameof(QueryName))]
-        public string QueryName { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the current operation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a driver load, including hash and signature information.
-    /// </summary>
-    /// <remarks>This class provides properties to access the hash values and signature details of a driver,
-    /// which can be used for validation and verification purposes.</remarks>
-    public sealed class DriverLoadMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the image loaded associated with the object.
-        /// </summary>
-        [JsonProperty(nameof(ImageLoaded))]
-        public string ImageLoaded { get; set; }
-
-        /// <summary>
-        /// Gets or sets the hash values associated with the object.
-        /// </summary>
-        [JsonProperty(nameof(Hashes))]
-        public string Hashes { get; set; }
-
-        /// <summary>
-        /// Gets or sets the digital signature associated with the object.
-        /// </summary>
-        [JsonProperty(nameof(Signature))]
-        public string Signature { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with an image load event in a process.
-    /// </summary>
-    /// <remarks>This class provides properties to access various details about the image load, such as the
-    /// company name, file version, and process ID. It is typically used in scenarios where detailed information about
-    /// loaded images is required for logging, auditing, or analysis purposes.</remarks>
-    public sealed class ImageLoadMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the name of the company associated with the entity.
-        /// </summary>
-        [JsonProperty(nameof(Company))]
-        public string Company { get; set; }
-
-        /// <summary>
-        /// Gets or sets the version of the file.
-        /// </summary>
-        [JsonProperty(nameof(FileVersion))]
-        public string FileVersion { get; set; }
-
-        /// <summary>
-        /// Gets or sets the hash values associated with the image load event.
-        /// </summary>
-        [JsonProperty(nameof(Hashes))]
-        public string Hashes { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the event.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the original file name of the image.
-        /// </summary>
-        [JsonProperty(nameof(OriginalFileName))]
-        public string OriginalFileName { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the image load event.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the product name associated with the image.
-        /// </summary>
-        [JsonProperty(nameof(Product))]
-        public string Product { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the image load event.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image loaded associated with the object.
-        /// </summary>
-        [JsonProperty(nameof(ImageLoaded))]
-        public string ImageLoaded { get; set; }
-    }
-
-    /// <summary>
-    /// Represents file create metadata associated with a file event, including process information and file details.
-    /// </summary>
-    /// <remarks>This class provides properties to access various details about a file event, such as the
-    /// process ID, the image name, the target filename, the creation time in UTC, and the user associated with the
-    /// event.</remarks>
-    public sealed class FileCreateMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the process ID associated with the file event.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the file event.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target filename involved in the file event.
-        /// </summary>
-        [JsonProperty(nameof(TargetFilename))]
-        public string TargetFilename { get; set; }
-
-        /// <summary>
-        /// Gets or sets the creation time of the file event in UTC.
-        /// </summary>
-        [JsonProperty(nameof(CreationUtcTime))]
-        public string CreationUtcTime { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user associated with the file event.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the raw access read permissions as a string.
-        /// </summary>
-        //[JsonProperty(nameof(RawAccessRead), NullValueHandling = NullValueHandling.Ignore)]
-        //public string RawAccessRead { get; set; }
-    }
-
-    public sealed class FileModifiedMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the process ID associated with the file modification event.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-        /// <summary>
-        /// Gets or sets the image name associated with the file modification event.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-        /// <summary>
-        /// Gets or sets the target filename that was modified.
-        /// </summary>
-        [JsonProperty(nameof(TargetFilename))]
-        public string TargetFilename { get; set; }
-        /// <summary>
-        /// Gets or sets the user who performed the file modification operation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the creation time of the file in UTC.
-        /// </summary>
-        [JsonProperty(nameof(CreationUtcTime), NullValueHandling = NullValueHandling.Ignore)]
-        public DateTime CreationUtcTime { get; set; }
-
-        /// <summary>
-        /// Gets or sets the previous creation time of the file in UTC.
-        /// </summary>
-        [JsonProperty(nameof(PreviousCreationUtcTime), NullValueHandling = NullValueHandling.Ignore)]
-        public DateTime PreviousCreationUtcTime { get; set; }
-    }
-
-
-    /// <summary>
-    /// Represents metadata for creating a stream hash, including details such as creation time, hash value, and
-    /// associated user.
-    /// </summary>
-    /// <remarks>This class is used to encapsulate metadata information required for the creation of a stream
-    /// hash. It includes properties for the creation time in UTC, the hash value, the image associated with the hash,
-    /// the process ID, the target filename, and the user who initiated the creation.</remarks>
-    public sealed class CreateStreamHashMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the creation time of the stream hash in UTC.
-        /// </summary>
-        [JsonProperty(nameof(CreationUtcTime))]
-        public string CreationUtcTime { get; set; }
-
-        /// <summary>
-        /// Gets or sets the hash value of the stream.
-        /// </summary>
-        [JsonProperty(nameof(Hash))]
-        public string Hash { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image associated with the stream hash.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the stream hash creation.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target filename for which the stream hash was created.
-        /// </summary>
-        [JsonProperty(nameof(TargetFilename))]
-        public string TargetFilename { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user who initiated the stream hash creation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a file deletion operation.
-    /// </summary>
-    /// <remarks>This class provides information about the file deletion process, including the image name,
-    /// process ID, target filename, and user responsible for the operation. It is intended to be used in scenarios
-    /// where tracking or auditing of file deletions is required.</remarks>
-    public sealed class FileDeleteMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the details of the file deletion operation.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the file deletion operation.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target filename that was deleted.
-        /// </summary>
-        [JsonProperty(nameof(TargetFilename))]
-        public string TargetFilename { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user who performed the file deletion operation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a registry addition event.
-    /// </summary>
-    /// <remarks>This class provides information about a registry addition, including the process ID,  the
-    /// image name, the target object, and the user responsible for the action.</remarks>
-    public sealed class RegistryAddMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the details of the registry addition operation.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the registry addition.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target object of the registry addition.
-        /// </summary>
-        [JsonProperty(nameof(TargetObject))]
-        public string TargetObject { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user who performed the registry addition operation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a registry delete operation.
-    /// </summary>
-    /// <remarks>This class provides details about a registry delete operation, including the image involved,
-    /// the process ID, and the target object of the operation. It is used to encapsulate information relevant to the
-    /// deletion event.</remarks>
-    public sealed class RegistryDeleteMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the details of the registry delete operation.
-        /// </summary>
-        [JsonProperty(nameof(Details))]
-        public string Details { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the registry delete operation.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the registry delete operation.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target object of the registry delete operation.
-        /// </summary>
-        [JsonProperty(nameof(TargetObject))]
-        public string TargetObject { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a registry set operation.
-    /// </summary>
-    /// <remarks>This class provides details about a registry set operation, including the user who performed
-    /// the operation, the target object affected, and additional descriptive information.</remarks>
-    public sealed class RegistrySetMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the details of the registry set operation.
-        /// </summary>
-        [JsonProperty(nameof(Details))]
-        public string Details { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the registry set operation.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the registry set operation.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target object of the registry set operation.
-        /// </summary>
-        [JsonProperty(nameof(TargetObject))]
-        public string TargetObject { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user who performed the registry set operation.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-    }
-
-    /// <summary>
-    /// Represents metadata associated with a registry rename event.
-    /// </summary>
-    /// <remarks>This class provides details about a registry rename event, including information about the process
-    /// and user involved.</remarks>
-    public sealed class RegistryRenameMetadata : IMetadata
-    {
-        /// <summary>
-        /// Gets or sets the details of the registry event.
-        /// </summary>
-        [JsonProperty(nameof(Details))]
-        public string Details { get; set; }
-
-        /// <summary>
-        /// Gets or sets the image name associated with the registry event.
-        /// </summary>
-        [JsonProperty(nameof(Image))]
-        public string Image { get; set; }
-
-        /// <summary>
-        /// Gets or sets the process ID associated with the registry event.
-        /// </summary>
-        [JsonProperty(nameof(ProcessId))]
-        public long ProcessId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the target object of the registry event.
-        /// </summary>
-        [JsonProperty(nameof(TargetObject))]
-        public string TargetObject { get; set; }
-
-        /// <summary>
-        /// Gets or sets the user who performed the registry event.
-        /// </summary>
-        [JsonProperty(nameof(User))]
-        public string User { get; set; }
-
-        /// <summary>
-        /// Gets or sets the new name of the registry key after the rename operation.
-        /// </summary>
-        [JsonProperty(nameof(NewName))]
-        public string NewName { get; set; }
-    }
 
     /// <summary>
     /// Provides functionality to decode Sysmon events from trace event data.
@@ -928,6 +113,8 @@ namespace Nanolite_agent.Helper
                 case FILEDELETE:
                 case FILEDELETEDETECTED:
                     return SysEventCode.FileDelete;
+                case RAWACCESSREAD:
+                    return SysEventCode.RawAccessReadDetected;
                 case FILECREATIONTIMECHANGED:
                     return SysEventCode.FileModified;
                 case FILECREATESTREAMHASH:
@@ -991,6 +178,9 @@ namespace Nanolite_agent.Helper
                     break;
                 case SysEventCode.FileModified:
                     metadataObj = SysmonEventDecoder.DecodeSysmonMetadata<FileModifiedMetadata>(origin);
+                    break;
+                case SysEventCode.RawAccessReadDetected:
+                    metadataObj = SysmonEventDecoder.DecodeSysmonMetadata<RawAccessReadMetadata>(origin);
                     break;
                 case SysEventCode.CreateStreamHash:
                     metadataObj = SysmonEventDecoder.DecodeSysmonMetadata<CreateStreamHashMetadata>(origin);
